frontend/requirements.txt
Flask==2.0.2
Flask-Minify==0.36
google-cloud-firestore==2.3.4
